Finally set up
Set up a blog template, will need to play with some files to make them for this project more specifically.
Got the app to work, played with some css stuff to make cards.
Firebase might be a headache, but if no need for pictures than its easy to avoid firebase and use mongoDB, will need to discuss.
TODO: CSS Styling to make the website less plain